+++ src/processor/detection.service.ts
@@ -1,6 +1,10 @@
 import * as ort from "onnxruntime-node";
 import { BaseDetectionService } from "../core/base-detection.service.js";
-import type { DebuggingOptions, DetectionOptions } from "../interface.js";
+import type {
+  DebuggingOptions,
+  DetectionOptions,
+  ProcessingEngine,
+} from "../interface.js";
 import { NodePlatformProvider } from "./platform.node.js";
 
 /**
@@ -11,7 +15,8 @@ export class DetectionService extends BaseDetectionService {
     session: ort.InferenceSession,
     options: Partial<DetectionOptions> = {},
     debugging: Partial<DebuggingOptions> = {},
+    engine: ProcessingEngine = "opencv",
   ) {
-    super(new NodePlatformProvider(), session, options, debugging);
+    super(new NodePlatformProvider(), session, options, debugging, engine);
   }
 }